% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{baker_map}
\alias{baker_map}
\title{Baker map}
\usage{
baker_map(x0, a)
}
\arguments{
\item{x0}{seed value}

\item{a}{parameter of the map range is greater than  0.5}
}
\value{
for 0<=x=<1/2 the map returns 2ax
for 1/2<=x<=1 the map returns a(2x-1)mod1
}
\description{
this is a chaotic map with the sensitive for the parameter value greater than 0.5
}
\examples{
baker_map(0.3,0.56)
}
