% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{logistic_map}
\alias{logistic_map}
\title{Logistic map}
\usage{
logistic_map(x0, a)
}
\arguments{
\item{x0}{the seed value range from 0 to 1}

\item{a}{the parameter ranging from 3.5 to 4}
}
\value{
the map returns the a*x(1-x) for input x
}
\description{
This is the most used chaotic map . The map is sensitive for the value of the parameter greater than 3.568
}
\examples{
logistic_map(0.26,3.5)



}
