% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{prbits}
\alias{prbits}
\title{Random Bit generator}
\usage{
prbits(n, Time = TRUE)
}
\arguments{
\item{n}{number of bits required}

\item{Time}{it is a boolean value of TRUE/FALSE if we want to generate time dependent random bits.i.e each time we call the function with same input different output will be generated.}
}
\value{
returns a vector of random bits of length n
}
\description{
this function generates random bits of desired length
}
\examples{
prbits(2)
prbits(2)
prbits(2,Time=FALSE)
prbits(2,Time=FALSE)
 prbits(10)

}
