% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{prcauchy}
\alias{prcauchy}
\title{Cauchy distribution}
\usage{
prcauchy(n, Time = TRUE)
}
\arguments{
\item{n}{How many numbers we want}

\item{Time}{time dependent or not}
}
\value{
a vector of n numbers from cauchy distribution
}
\description{
This function generates random numbers from standard  cauchy distribution
}
\examples{
prcauchy(10)
prcauchy(10,Time=TRUE)
prcauchy(10,Time=TRUE)
}
