% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{prexp}
\alias{prexp}
\title{Exponentail distribution}
\usage{
prexp(n, Time = TRUE)
}
\arguments{
\item{n}{how many numbers we need}

\item{Time}{time dependent or not}
}
\value{
a vector of n numbers from exponential distribution
}
\description{
This function generates random numbers from exponentail distribution
}
\examples{
 prexp(10)
   prexp(10)
   prexp(10,FALSE)
   prexp(10,FALSE)
}
