% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNGCorrNor.R
\name{PRNGCorrNor}
\alias{PRNGCorrNor}
\title{Pseudo-Random Number Generation of Correlated Normal Outcomes}
\usage{
PRNGCorrNor(seed, S, Tvec, rhoyy, rhoxy, rhoyx, TDCTypes, dataTypes, beta, pred,
  sigma)
}
\arguments{
\item{seed}{The seed used to control pseudo-random number generation.}

\item{S}{The number of subjects or groups to be generated.  These are assumed to be collections of auto-correlated data.}

\item{Tvec}{The vector indicating the number of times (for each subject) or the number of units (for each group) to be generated.}

\item{rhoyy}{The auto-regressive weight, a correlation between responses from the same subject or group.}

\item{rhoxy}{The weight representing the correlation between "previous" time-dependent covariate values on response values.  This is associated with Type II time-dependent covariates.}

\item{rhoyx}{The weight representing the correlation between "previous" responses and time-dependent covariate values.  This is associated with Type IV time-dependent covariates.}

\item{TDCTypes}{The vector indicating the type (I, II, III, or IV) of each time-dependent covariate.}

\item{dataTypes}{The vector indicating the type of each predictor.  Either 'b' for binary or 'c' for continuous.}

\item{beta}{The vector of "true" parameter values according to which data will be randomly generated.  This includes coefficients for all time-independent covariates, all time-dependent covariates, and an intercept.}

\item{pred}{A vector of parameters describing the distributions of predictors.  Each binary predictor should have an associated probability of success, while each continuous predictors should have an associated standard deviation.}

\item{sigma}{The error standard deviation associated with the normal outcomes generated.}
}
\description{
This function produces pseudo-random numbers for both predictors and responses for correlated (longitudinal) data.  The responses are generated as normal data.  The predictors can be either continuous or binary, and can include both time-independent and time-dependent covariates.  Time-dependent covariates can be controlled as either Type I, Type II, Type III, or Type IV time-dependent covariates.  The function returns a response vector, a design matric for the time-dependent covariates (with intercept column), and a design matrix for the time-independent covariates.
}
\examples{
PRNGCorrNor()
}
\keyword{Correlated}
\keyword{Data}
\keyword{Generation}
\keyword{Simulation}
