% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNGODBinBeta.R
\name{PRNGODBinBeta}
\alias{PRNGODBinBeta}
\title{Pseudo-Random Number Generation of Overdispersed Correlated Binary (Bernoulli-Beta) or Count (Binomial-Beta) Outcomes}
\usage{
PRNGODBinBeta(M, mvec, D, beta0, beta1, x, seed)
}
\arguments{
\item{M}{The number of groups in the resulting data.}

\item{mvec}{A vector indicating the length of each group in the data.}

\item{D}{A vector of denominators for each response; a vector of 1's indicates Bernoulli data.}

\item{beta0}{The true intercept in the systematic component of the model.}

\item{beta1}{The true slope in the systematic component of the model.}

\item{x}{A vector of predictor values.}

\item{seed}{The seed for data generation.}
}
\description{
This function produces pseudo-random numbers for responses for correlated (longitudinal) data.  The responses are generated as either binary (Bernoulli) data or count (Binomial), using a transformed Beta random effect to induce autocorrelation.  Once the Beta-distributed random effects are generated, the values are transformed according to a logit transformation to ensure they interact with the response similarly to otherpredictors.  These random transformed effects are then passed through the systematic component to a Binomial-Beta pseudo-random number generator from the VGAM package.  Therefore there are two instances of random effects: in the transformed Beta random effects, and through the Binomial-Beta pseudo-random number generation.  The predictors must be provided as a vector.  The function returns a response vector, labeled "Outcomes".
}
\examples{
PRNGODBinBeta()
}
\keyword{Correlated}
\keyword{Data}
\keyword{Generation}
\keyword{Simulation}
