% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNGODBinNormal.R
\name{PRNGODBinNormal}
\alias{PRNGODBinNormal}
\title{Pseudo-Random Number Generation of Overdispersed Correlated Binary (Bernoulli-Normal) or Count (Binomial-Normal) Outcomes}
\usage{
PRNGODBinNormal(M, mvec, mu_r, s_r, D, beta0, beta1, x, seed)
}
\arguments{
\item{M}{The number of groups in the resulting data.}

\item{mvec}{A vector indicating the length of each group in the data.}

\item{mu_r}{The mean of the normal random effect.}

\item{s_r}{The standard deviation of the normal random effect.}

\item{D}{A vector of denominators for each response; a vector of 1's indicates Bernoulli data.}

\item{beta0}{The true intercept in the systematic component of the model.}

\item{beta1}{The true slope in the systematic component of the model.}

\item{x}{A vector of predictor values.}

\item{seed}{The seed for data generation.}
}
\description{
This function produces pseudo-random numbers for responses for correlated (longitudinal) data.  The responses are generated as either binary (Bernoulli) data or count (Binomial), using a normal random effect to induce autocorrelation.  These random effects are then passed through the systematic component to a Binomial-Beta pseudo-random number generator from the VGAM package.  Therefore there are two instances of random effects: in the normal random effects, and through the Binomial-Beta pseudo-random number generation.  The predictors must be provided as a vector.  The function returns a response vector, labeled "Outcomes".
}
\examples{
PRNGODBinNormal()
}
\keyword{Correlated}
\keyword{Data}
\keyword{Generation}
\keyword{Simulation}
