% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{rcauchy}
\alias{rcauchy}
\title{Cauchy distribution}
\usage{
rcauchy(n, Time = TRUE)
}
\arguments{
\item{n}{How many numbers we want}

\item{Time}{time dependent or not}
}
\value{
a vector of n numbers from cauchy distribution
}
\description{
This function generates random numbers from standard  cauchy distribution
}
\examples{
rcauchy(10)
rcauchy(10,Time=TRUE)
rcauchy(10,Time=TRUE)
}
