% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{rexp}
\alias{rexp}
\title{Exponentail distribution}
\usage{
rexp(n, Time = TRUE)
}
\arguments{
\item{n}{how many numbers we need}

\item{Time}{time dependent or not}
}
\value{
a vector of n numbers from exponential distribution
}
\description{
This function generates random numbers from exponentail distribution
}
\examples{
 rexp(10)
   rexp(10)
   rexp(10,FALSE)
   rexp(10,FALSE)
}
