% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{rnorm}
\alias{rnorm}
\title{Generating numbers form    Normal distribution
here we use  Box Muler transform to obtain normal random variable}
\usage{
rnorm(n)
}
\arguments{
\item{n}{number required}
}
\value{
a list of pseudo random numbers from normal distribution
}
\description{
Generating numbers form    Normal distribution
here we use  Box Muler transform to obtain normal random variable
}
\examples{
rnorm(10)
 rnorm(100)

}
