% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{runf}
\alias{runf}
\title{Uniformly  Pseudo random  number generator}
\usage{
runf(
  N = 100,
  Time = TRUE,
  n0 = 5,
  x00 = 0.5362,
  x01 = 0.357,
  x02 = 0.235,
  a1 = 3.69,
  a2 = 0.7
)
}
\arguments{
\item{N}{How many  numbers  are required}

\item{Time}{if enabled TRUE the numbers are time dependent}

\item{n0}{seed value of linear congruence map   it can take value of any  natural number}

\item{x00}{seed value of saw-tooth map values from 0 to 1}

\item{x01}{seed value of logistic map values from 0 to 1}

\item{x02}{seed value of baker map}

\item{a1}{parameter of logistic map the value takes from 3.5 to 4}

\item{a2}{parameter of baker map the value it takes values greater than or equalt to 0.5}
}
\value{
gives a vector  of pseudo random numbers generated of desired length
}
\description{
this function generates random numbers which follow uniform distribution \code{[0,1]}
}
\examples{
  runf(10)
runf(10,Time=TRUE)
runf(10,Time=TRUE)
runf(10,Time=TRUE)
runf(10,2)
runf(10,Time=TRUE,2)
runf(10,Time=TRUE,2)

runf(10,5,0.52)
 runf(15,2,0.352)

  runf(10,2,0.652,0.235)
   runf(10,Time=TRUE,2,0.652,0.235)
  runf(9,7,0.52,0.4235,0.389)
  runf(10,Time=TRUE,2,0.752,0.235,0.351,3.8)
}
