% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{saw_tooth}
\alias{saw_tooth}
\title{Saw tooth map}
\usage{
saw_tooth(x0)
}
\arguments{
\item{x0}{seed value ranging from 0 to 1}
}
\value{
(3*x) mod(1)
}
\description{
saw tooth map is a family of maps as f(x)=b*x mod1
}
\examples{
saw_tooth(0.6)
}
