% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRNG.R
\name{time}
\alias{time}
\title{time function}
\usage{
time()
}
\value{
t fractional value of the time
}
\description{
This function is used to generate a time of the system to be used for generating time dependent random numbers   precise upto micro-seconds
}
\examples{
time()
}
